+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.5
 
+* Add `selectT`, see #42.
 * Add more general instances for `IdentityT` and `ReaderT`. This is technically
   a breaking change because `Selective` is not a superclass of `Monad`. See #38.